refactor(niuniu): 重构用户状态和 Buff 相关逻辑
- 优化 UserState 类的 set_or_get 方法，提高代码可读性和性能
- 重构 get_buffs 函数，确保返回值类型一致
- 调整 PropModel 中的字段类型，提高精度
- 优化击剑和打胶相关逻辑，提高代码可维护性
- 移除未使用的导入和优化导入结构，提高代码整洁度

+++ niuniu/config_loader.py
@@ -121,7 +121,7 @@ class PropModel(BaseModel):
     """打胶效果的倍数"""
     glue_negative_weight: float = 1
     """打胶触发负面事件的倍数"""
-    expire_time: int = 0
+    expire_time: float = 0
     """过期时间"""
